Card display issue when sensor has no values
When the delijn sensor has no values (no passages) in HA, the card will not display at all when used in a vertical stack.
Ended up completely rewriting the code due to some other issues like the last passage being stuck on the card until the next morning and added a new feature to show the delay in minutes.